DCMAW-5594: Rename env to config
