+++ deploy/scripts/src/mobile/utils/config.ts
