+++ deploy/scripts/src/mobile/utils/url.ts
@@ -1,5 +1,5 @@
 import { URL } from './k6/url'
-import { env } from './env'
+import { env } from './config'
 
 const OAUTH_ROUTE = '/dca/oauth2'
